Cleanup umem usage, add compile switches for using u64 or usize for umem, while also supporting u128 over an unexposed feature

+++ memflow-win32/src/kernel/start_block/aarch64.rs
@@ -11,7 +11,7 @@ pub const PHYS_BASE: umem = mem::gb(1);
 // mem here has to be a single page (4kb sized)
 fn find_pt(addr: Address, mem: &[u8]) -> Option<Address> {
     // TODO: global define / config setting
-    let max_mem = mem::gb(512);
+    let max_mem = mem::gb(512) as u64;
 
     let pte = u64::from_le_bytes(mem[0..8].try_into().unwrap());
 
@@ -24,7 +24,7 @@ fn find_pt(addr: Address, mem: &[u8]) -> Option<Address> {
     mem[0x800..]
         .chunks(8)
         .map(|c| u64::from_le_bytes(c.try_into().unwrap()))
-        .find(|a| (a ^ 0xf03) & (!0u64 >> 12) == addr.to_umem())?;
+        .find(|a| (a ^ 0xf03) & (!0u64 >> 12) == addr.to_umem() as u64)?;
 
     // A page table does need to have some entries, right? Particularly, kernel-side page table
     // entries must exist
